Changed java type for IntegerEditingSupport from Integer to Long

+++ plugins/org.yakindu.sct.simulation.ui/src/org/yakindu/sct/simulation/ui/view/editing/IntegerEditingSupport.java
@@ -36,7 +36,7 @@ public class IntegerEditingSupport extends ScopeSlotEditingSupport {
 		textCellEditor.setValidator(new ICellEditorValidator() {
 			public String isValid(Object value) {
 				try {
-					Integer.parseInt((String) value);
+					Long.parseLong((String) value);
 				} catch (NumberFormatException e) {
 					return "No valid integer value!";
 				}
@@ -55,7 +55,7 @@ public class IntegerEditingSupport extends ScopeSlotEditingSupport {
 	@Override
 	protected Object convertValue(Object element, Object value) {
 		if (value != null && value instanceof String && ((String) value).trim().length() > 0) {
-			return Integer.parseInt((String) value);
+			return Long.parseLong((String) value);
 		}
 		return null;
 	}
